test: split slow test file to increase parallelism
```
  Slow test file: [chromium] › dataset-tag-persistence.spec.ts (52.4s)
  Consider splitting slow test files to speed up parallel execution
```
